Addition of data analysis app for physical therapy (app2001)
This update adds a new app (app2001) which is the initial interactive
app for analyzing BVH files to extract data needed for physical therapy
project.  The new app is interactive in that it reads its instructions
from a command file (skacommands,txt) rather than having them written
into program code.

+++ SKA/apps/app1001/AnimationControl.h
@@ -33,7 +33,7 @@ private:
 	vector<MotionDataSpec> specs;
 
 public:
-	short size() { return specs.size(); }
+	short size() { return (short)specs.size(); }
 	void addSpec(string _seqID, string _BVH_file, string _quat_file)
 	{
 		specs.push_back(MotionDataSpec(_seqID, _BVH_file, _quat_file));
